Simplify VLP-16 packet parsing
The reason for this change is to enable an easy way to produce mock
VLP-16 serialized data in tests, using a single call to binary.Write.

However, this change also ensures that the Packet struct models
byte-for-byte an on-the-wire VLP-16 packet, which simplifies parsing to
a single binary.Read call.